RM-895: Ensure RM actions with side effects don't try and execute if things are frozen RM-965: Unable to declare a record from the repository view.
git-svn-id: https://svn.alfresco.com/repos/alfresco-enterprise/modules/recordsmanagement/HEAD@55371 c4b6b30b-aa2e-2d43-bbcb-ca4b014f7261

+++ rm-server/source/java/org/alfresco/module/org_alfresco_module_rm/action/impl/UndeclareRecordAction.java
@@ -44,20 +44,25 @@ public class UndeclareRecordAction extends RMActionExecuterAbstractBase
     @Override
     protected void executeImpl(Action action, NodeRef actionedUponNodeRef)
     {
-        if (recordService.isRecord(actionedUponNodeRef) == true)
-        {
-            if (recordService.isDeclared(actionedUponNodeRef) == true)
-            {
-                // Remove the declared aspect
-                this.nodeService.removeAspect(actionedUponNodeRef, ASPECT_DECLARED_RECORD);
-            }
-        }
-        else
-        {
-            if (logger.isWarnEnabled())
-            {
-                logger.warn(I18NUtil.getMessage(MSG_RECORDS_ONLY_UNDECLARED));
-            }
-        }
+    	if (nodeService.exists(actionedUponNodeRef) == true)
+    	{
+	        if (recordService.isRecord(actionedUponNodeRef) == true)
+	        {
+	        	// repoen if already complete and not frozen
+	            if (recordService.isDeclared(actionedUponNodeRef) == true &&
+	                freezeService.isFrozen(actionedUponNodeRef) == false)
+	            {
+	                // Remove the declared aspect
+	                this.nodeService.removeAspect(actionedUponNodeRef, ASPECT_DECLARED_RECORD);
+	            }
+	        }
+	        else
+	        {
+	            if (logger.isWarnEnabled())
+	            {
+	                logger.warn(I18NUtil.getMessage(MSG_RECORDS_ONLY_UNDECLARED));
+	            }
+	        }
+    	}
     } 
 }
